Move _http to utils.py
