+++ flask_openapi3/view.py
@@ -5,11 +5,11 @@ import typing
 from copy import deepcopy
 from typing import Optional, List, Dict, Any, Callable
 
-from ._http import HTTPMethod
 from .models import ExternalDocumentation
 from .models import Server
 from .models import Tag
 from .types import ResponseDict
+from .utils import HTTPMethod
 from .utils import convert_responses_key_to_string
 from .utils import get_operation
 from .utils import get_operation_id_for_path
